refactor: extract combat projection module
